Port Android Studio 2.3.3
Port Android Studio 2.3.3
Change SDK version to 26
Change package to kr.ac.pusan.walkover.autotrackingcctv
