--- app/src/androidTest/java/org/androidtown/auto_tracking_cctvapp/ApplicationTest.java
@@ -1,13 +1,0 @@
-package org.androidtown.auto_tracking_cctvapp;
-
-import android.app.Application;
-import android.test.ApplicationTestCase;
-
-/**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
- */
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
-        super(Application.class);
-    }
-}
