+++ app/src/main/AndroidManifest.xml
@@ -1,14 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.androidtown.auto_tracking_cctvapp">
+    package="kr.ac.pusan.walkover.autotrackingcctv">
+
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="android.support.multidex.MultiDexApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+
+        <activity android:name=".MainActivity" />
         <activity android:name=".MainActivity_login">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -16,9 +20,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity android:name=".camera_select"></activity>
-        <activity android:name=".selected_camera_play"></activity>
+        <activity android:name=".camera_select" />
+        <activity android:name=".selected_camera_play" />
 
         <service android:name=".fcm.MyFirebaseMessagingService">
             <intent-filter>
@@ -31,30 +34,6 @@
                 <action android:name="com.google.firebase.INSTANCE_ID_EVENT" />
             </intent-filter>
         </service>
-
     </application>
-
-
-    <!--<service-->
-    <!--android:name=".MyFirebaseInstanceIDService">-->
-    <!--<intent-filter>-->
-    <!--<action android:name="com.google.firebase.INSTANCE_ID_EVENT"/>-->
-    <!--</intent-filter>-->
-    <!--</service>-->
-
-    <!--&lt;!&ndash; Set custom default icon. This is used when no icon is set for incoming notification messages.-->
-    <!--See README(https://goo.gl/l4GJaQ) for more. &ndash;&gt;-->
-    <!--<meta-data-->
-    <!--android:name="com.google.firebase.messaging.default_notification_icon"-->
-    <!--android:resource="@drawable/ic_stat_ic_notification" />-->
-    <!--&lt;!&ndash; Set color used with incoming notification messages. This is used when no color is set for the incoming-->
-    <!--notification message. See README(https://goo.gl/6BKBk7) for more. &ndash;&gt;-->
-    <!--<meta-data-->
-    <!--android:name="com.google.firebase.messaging.default_notification_color"-->
-    <!--android:resource="@color/colorAccent" />-->
-
-
-    <!-- for internet communication -->
-    <uses-permission android:name="android.permission.INTERNET" />
 
 </manifest>
