+++ app/src/main/java/kr/ac/pusan/walkover/autotrackingcctv/MainActivity.java
@@ -1,0 +1,13 @@
+package kr.ac.pusan.walkover.autotrackingcctv;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
