+++ app/src/main/java/kr/ac/pusan/walkover/autotrackingcctv/retrofit/Direction.java
@@ -1,0 +1,14 @@
+package kr.ac.pusan.walkover.autotrackingcctv.retrofit;
+
+/**
+ * Created by soeun on 2017-07-24.
+ */
+public class Direction {
+    private Integer camera_num;
+    private String direction;
+
+    public Direction(Integer camera_num, String direction) {
+        this.camera_num = camera_num;
+        this.direction = direction;
+    }
+}
