+++ app/src/main/java/kr/ac/pusan/walkover/autotrackingcctv/retrofit/Mode.java
@@ -1,0 +1,14 @@
+package kr.ac.pusan.walkover.autotrackingcctv.retrofit;
+
+/**
+ * Created by soeun on 2017-08-01.
+ */
+public class Mode {
+    private Integer camera_num;
+    private String mode;
+
+    public Mode(Integer camera_num, String mode) {
+        this.camera_num = camera_num;
+        this.mode = mode;
+    }
+}
