+++ app/src/main/java/kr/ac/pusan/walkover/autotrackingcctv/retrofit/direction_message.java
@@ -1,0 +1,10 @@
+package kr.ac.pusan.walkover.autotrackingcctv.retrofit;
+
+/**
+ * Created by soeun on 2017-07-17.
+ */
+public class direction_message {
+
+    private String direction_message;
+
+}
