+++ app/src/main/java/kr/ac/pusan/walkover/autotrackingcctv/retrofit/mode_message.java
@@ -1,0 +1,8 @@
+package kr.ac.pusan.walkover.autotrackingcctv.retrofit;
+
+/**
+ * Created by soeun on 2017-08-01.
+ */
+public class mode_message {
+    private String mode_message;
+}
