--- app/src/main/java/org/androidtown/auto_tracking_cctvapp/fcm/MyFirebaseInstanceIdService.java
@@ -1,49 +1,0 @@
-package org.androidtown.auto_tracking_cctvapp.fcm;
-
-import android.util.Log;
-
-import com.google.firebase.iid.FirebaseInstanceId;
-import com.google.firebase.iid.FirebaseInstanceIdService;
-
-import java.io.IOException;
-
-import okhttp3.FormBody;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.RequestBody;
-
-/**
- * Created by soeun on 2017-08-17.
- */
-public class MyFirebaseInstanceIdService extends FirebaseInstanceIdService {
-    private final static String TAG = "MyFirebaseIIdService";
-
-    @Override
-    public void onTokenRefresh() {
-//        super.onTokenRefresh();
-        String token = FirebaseInstanceId.getInstance().getToken();
-        Log.d(TAG, "Refreshed token: " + token);
-
-        sendRegistrationToServer(token);
-    }
-
-    private void sendRegistrationToServer(String token) {
-        OkHttpClient client = new OkHttpClient();
-        RequestBody body = new FormBody.Builder()
-                .add("Token", token)
-                .build();
-
-        //request
-        Request request = new Request.Builder()
-                .url("http://164.125.68.74/fcm/register") //수정
-                .post(body)
-                .build();
-
-        try {
-            client.newCall(request).execute();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-}
