--- app/src/main/java/org/androidtown/auto_tracking_cctvapp/retrofit/Direction.java
@@ -1,14 +1,0 @@
-package org.androidtown.auto_tracking_cctvapp.retrofit;
-
-/**
- * Created by soeun on 2017-07-24.
- */
-public class Direction {
-    private Integer camera_num;
-    private String direction;
-
-    public Direction(Integer camera_num, String direction) {
-        this.camera_num = camera_num;
-        this.direction = direction;
-    }
-}
