--- app/src/main/java/org/androidtown/auto_tracking_cctvapp/retrofit/Mode.java
@@ -1,14 +1,0 @@
-package org.androidtown.auto_tracking_cctvapp.retrofit;
-
-/**
- * Created by soeun on 2017-08-01.
- */
-public class Mode {
-    private Integer camera_num;
-    private String mode;
-
-    public Mode(Integer camera_num, String mode) {
-        this.camera_num = camera_num;
-        this.mode = mode;
-    }
-}
