--- app/src/main/java/org/androidtown/auto_tracking_cctvapp/retrofit/direction_message.java
@@ -1,10 +1,0 @@
-package org.androidtown.auto_tracking_cctvapp.retrofit;
-
-/**
- * Created by soeun on 2017-07-17.
- */
-public class direction_message {
-
-    private String direction_message;
-
-}
