--- app/src/main/java/org/androidtown/auto_tracking_cctvapp/retrofit/mode_message.java
@@ -1,8 +1,0 @@
-package org.androidtown.auto_tracking_cctvapp.retrofit;
-
-/**
- * Created by soeun on 2017-08-01.
- */
-public class mode_message {
-    private String mode_message;
-}
