+++ app/src/main/res/layout/activity_main_server_connect.xml
@@ -3,16 +3,16 @@
 <TableLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="600px"
     android:layout_height="wrap_content"
-    android:stretchColumns="1"
-    android:layout_gravity="center">
+    android:layout_gravity="center"
+    android:stretchColumns="1">
 
     <TableRow>
 
         <TextView
-            android:text="서버 접속"
             android:layout_span="2"
-            android:textSize="10pt"
-            android:textColor="#ffff7825" />
+            android:text="서버 접속"
+            android:textColor="#ffff7825"
+            android:textSize="10pt" />
     </TableRow>
 
     <TableRow>
@@ -21,9 +21,9 @@
 
         <EditText
             android:id="@+id/server_ip_address"
-            android:inputType="number"
             android:digits="1234567890."
             android:hint="164.125.68.74"
+            android:inputType="number"
             android:text="164.125.68.74" />
     </TableRow>
 
@@ -33,8 +33,8 @@
 
         <EditText
             android:id="@+id/server_port_num"
-            android:inputType="number"
             android:hint="13000"
+            android:inputType="number"
             android:text="13000" />
     </TableRow>
 
@@ -42,8 +42,8 @@
 
         <Button
             android:id="@+id/server_connect_btn"
-            android:text="Connect"
             android:layout_column="0"
-            android:layout_span="2" />
+            android:layout_span="2"
+            android:text="Connect" />
     </TableRow>
 </TableLayout>
