+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">"Auto-tracking-cctv app "</string>
+    <string name="app_name">AutoTrackingCCTV</string>
 </resources>
