+++ app/src/test/java/kr/ac/pusan/walkover/autotrackingcctv/ExampleUnitTest.java
@@ -1,0 +1,17 @@
+package kr.ac.pusan.walkover.autotrackingcctv;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+public class ExampleUnitTest {
+    @Test
+    public void addition_isCorrect() throws Exception {
+        assertEquals(4, 2 + 2);
+    }
+}
